refactor: 카테고리 코드 리팩토링 (#114)
