+++ src/main/java/ku_rum/backend/global/exception/handler/CategoryExceptionControllerAdvice.java
@@ -1,17 +1,17 @@
-package ku_rum.backend.global.handler;
+package ku_rum.backend.global.exception.handler;
 
 import jakarta.annotation.Priority;
 import ku_rum.backend.global.exception.category.CategoryNotExistException;
 import ku_rum.backend.global.exception.category.CategoryNotProvidingDetailException;
-import ku_rum.backend.global.response.BaseErrorResponse;
+import ku_rum.backend.global.support.response.BaseErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import static ku_rum.backend.global.response.status.BaseExceptionResponseStatus.CATEGORYNAME_NOT_PROVIDING_DETAIL;
-import static ku_rum.backend.global.response.status.BaseExceptionResponseStatus.CATEGORY_NAME_NOT_EXIST;
+import static ku_rum.backend.global.support.response.status.BaseExceptionResponseStatus.CATEGORYNAME_NOT_PROVIDING_DETAIL;
+import static ku_rum.backend.global.support.response.status.BaseExceptionResponseStatus.CATEGORY_NAME_NOT_EXIST;
 
 @Slf4j
 @Priority(0)
